refactor: rename creation_functions to creators
