+++ ml/postprocessing/mediapipe_hand_detection.py
@@ -3,7 +3,7 @@ import numpy as np
 import cv2
 
 from .utils import generate_anchors_and_decode
-from ..messages.creation_functions import create_detection_message
+from ..messages.creators import create_detection_message
 
 
 class MPHandDetectionParser(dai.node.ThreadedHostNode):
