+++ ml/postprocessing/mediapipe_hand_landmarker.py
@@ -1,7 +1,7 @@
 import depthai as dai
 import numpy as np
 
-from ..messages.creation_functions import create_hand_keypoints_message
+from ..messages.creators import create_hand_keypoints_message
 
 
 class MPHandLandmarkParser(dai.node.ThreadedHostNode):
